packages/agenthub-iot-python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agenthub-iot"
version = "1.0.0"
description = "AgentHub IoT SDK for Python - Build autonomous IoT agents on Avalanche"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "AgentHub Protocol", email = "info@agenthub.protocol"}
]
keywords = ["agenthub", "iot", "avalanche", "blockchain", "x402", "raspberry-pi"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "web3>=6.0.0",
    "eth-account>=0.8.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/agenthub/agenthub-iot-python"
Documentation = "https://docs.agenthub.protocol"
Repository = "https://github.com/agenthub/agenthub-iot-python.git"
Issues = "https://github.com/agenthub/agenthub-iot-python/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
agenthub_iot = ["py.typed"]

